moving test files around and fixing eslint errors to fix npm run build command
