+++ package.json
@@ -1,11 +1,11 @@
 {
-  "name": "brauhausts",
+  "name": "@alchematic/brauhausts",
   "version": "1.0.0",
   "description": "A typescript fork of the brauhausjs library",
   "main": "build/main/index.js",
   "typings": "build/main/index.d.ts",
   "module": "build/module/index.js",
-  "repository": "https://github.com/worbunkel/brauhausts",
+  "repository": "https://github.com/alchematic/brauhausts",
   "license": "MIT",
   "keywords": [],
   "scripts": {
