+++ src/lib/recipe-timeline.spec.ts
@@ -1,5 +1,5 @@
-import testRecipe2 from '../../testJsonFiles/test-recipe-2.json';
-import testRecipe3 from '../../testJsonFiles/test-recipe-3.json';
+import testRecipe2 from '../test-recipe-2.json';
+import testRecipe3 from '../test-recipe-3.json';
 import { calculateRecipe, createRecipe, Recipe } from './recipe';
 import { computeRecipeTimeline } from './recipe-timeline';
